Move ghrsstl4 driver to drivers directory
